+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/flow/action/BalanceNotificationAction.java
@@ -1,0 +1,6 @@
+package com.ssafy.shinhanflow.flow.action;
+
+public class BalanceNotificationAction {
+	Long userId;
+	String 
+}
