+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/flow/action/TextNotificationAction.java
@@ -1,0 +1,4 @@
+package com.ssafy.shinhanflow.flow.action;
+
+public class TextNotificationAction {
+}
