+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/flow/trigger/Trigger.java
@@ -1,4 +1,4 @@
-package com.ssafy.shinhanflow.trigger;
+package com.ssafy.shinhanflow.flow.trigger;
 
 /**
  * Trigger interface
