+++ be/ShinhanFlow/src/main/java/com/ssafy/shinhanflow/flow/trigger/date/MultiDateTrigger.java
@@ -1,10 +1,9 @@
-package com.ssafy.shinhanflow.trigger.date;
+package com.ssafy.shinhanflow.flow.trigger.date;
 
 import java.time.LocalDate;
-import java.util.List;
 import java.util.Set;
 
-import com.ssafy.shinhanflow.trigger.Trigger;
+import com.ssafy.shinhanflow.flow.trigger.Trigger;
 
 import lombok.NonNull;
